requirements/base.txt
# SHA1:a9dde048f1ee1f00586264d726d0e89f16e56183
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-e file:.
    # via -r requirements/base.in
asgiref==3.8.1
    # via django
django==4.2.17
    # via
    #   djangorestframework
    #   expense-counter
djangorestframework==3.15.2
    # via expense-counter
psycopg2-binary==2.9.5
    # via expense-counter
python-dateutil==2.9.0.post0
    # via expense-counter
python-dotenv==1.0.1
    # via expense-counter
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via asgiref
